Renaming of File and folder
Renamed the "JSON" folder to "Data"
